Add logic of request processing
In separated cladd RequestProcessor was implemented logic of processing
different types of messages
